refactor(render): phase8 mesh uses shared uniform factory
