+++ view/render/src/lib.rs
@@ -9,6 +9,7 @@ pub mod render_3d;
 pub mod shader;
 pub mod surface;
 pub mod toolpath;
+pub mod uniform_factory;
 pub mod vertex_2d;
 pub mod vertex_3d;
 pub mod wireframe;
